day_46/pyproject.toml
[tool.poetry]
name = "day_46"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
python-dotenv = "^0.17.1"
spotipy = "^2.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
